+++ submission/pipelines/generate_pipelines.py
@@ -7,13 +7,12 @@ from d3m.metadata import (
     base as metadata_base, pipeline as pipeline_module
 )
 
-from byudml import __imputer_path__
 from byudml.imputer.random_sampling_imputer import RandomSamplingImputer
 from byudml.metafeature_extraction.metafeature_extraction import MetafeatureExtractor
 from byudml import __imputer_version__, __imputer_path__,  __metafeature_version__,  __metafeature_path__
 import sys
 sys.path.append('./submission')
-from utils import get_new_d3m_path, clear_directory, write_pipeline_for_submission
+from utils import get_new_d3m_path, clear_directory, write_pipeline_for_submission, get
 
 real_mongo_port=12345
 lab_hostname = "computer"
@@ -405,12 +404,11 @@ def remove_digests(
         if step['primitive']['id'] not in exclude_primitives:
             del step['primitive']['digest']
         
-
     return pipeline_json_structure
 
 
 def update_pipeline(
-    pipeline_json_structure, filename=None
+    pipeline_to_update, filename=None
 ):
     """
     This function updates the pipeline's digests and version numbers
@@ -422,19 +420,19 @@ def update_pipeline(
 
     :return a pipeline with updated digests
     """
-    if pipeline_json_structure is None and filename is None:
+    if pipeline_to_update is None and filename is None:
         raise ValueError("No pipeline json was given")
-    elif pipeline_json_structure is None:
+    elif pipeline_to_update is None:
         with open(filename, "r") as file:
             # NOTE: must be a pipeline with no digests, or recent digests
             # NOTE: reading this in as straight JSON doesn't work so we have to use the pipeline_module
-            pipeline_json_structure = pipeline_module.Pipeline.from_json(string_or_file=file).to_json_structure()
+            pipeline_to_update = pipeline_module.Pipeline.from_json(string_or_file=file).to_json_structure()
     else:
         try:
-            pipeline_json_structure = pipeline_module.Pipeline.from_json(json.dumps(pipeline_json_structure)).to_json_structure()
+            pipeline_to_update = pipeline_module.Pipeline.from_json(json.dumps(pipeline_to_update)).to_json_structure()
         except Exception as e:
             pass
-    for step in pipeline_json_structure['steps']:
+    for step in pipeline_to_update['steps']:
         # if not updated, check and update
         primitive = pipeline_module.PrimitiveStep(
             primitive=d3m_index.get_primitive(
@@ -450,30 +448,7 @@ def update_pipeline(
         version_matches = check_step["primitive"]["version"] == step["primitive"]["version"]
         if not version_matches:
             step["primitive"]["version"] = check_step["primitive"]["version"]
-        # make sure the digest exists
-        digests_match = "digest" in step["primitive"] and check_step["primitive"]["digest"] == step["primitive"]["digest"]
-        if not digests_match:
-            step["primitive"]["digest"] = check_step["primitive"]["digest"]
-
-    return pipeline_json_structure
-
-def get_pipeline_from_database(pipeline_id, mongo_client):
-    """
-    This function gets a pipeline from our local database given an id
-
-    Parameters
-    ----------
-    pipeline_id: the id of the pipeline to grab
-    mongo_client: a connection to the database
-
-    :return a pipeline matching the id
-    """
-    collection = mongo_client.metalearning.pipelines
-    pipeline_to_write = collection.find({"id": pipeline_id})
-    for pipeline in pipeline_to_write:
-        # should only be one pipeline
-        return pipeline
-
+        
 def add_best_pipelines():
     """
     This function checks the best_pipelines.csv for the best pipelines for a dataset, prepares and updates it, and writes it to the submodule.
